Moved logic of publishing proposal to cwd-proposal-single to a hook so that we can publish proposals from anywhere. Also moved simulate msg function to a hook.
