+++ packages/stateful/hooks/index.ts
@@ -1,11 +1,13 @@
 export * from './contracts'
 
 export * from './useAwaitNextBlock'
+export * from './useCwdProposalSinglePublishProposal'
 export * from './useEncodedCwdProposalSinglePrefill'
 export * from './useInbox'
 export * from './useLoadingDaoCardInfos'
 export * from './usePinnedDaos'
 export * from './useProfile'
+export * from './useSimulateCosmosMsgs'
 export * from './useUpdateNavigatingHref'
 export * from './useVotingModule'
 export * from './useWalletProfile'
