+++ packages/stateful/package.json
@@ -27,6 +27,7 @@
     "file-saver": "^2.0.5",
     "fuse.js": "^6.6.2",
     "i18next": "^21.8.10",
+    "interchain-rpc": "^1.5.0",
     "json2csv": "^5.0.7",
     "json5": "^2.2.0",
     "lodash.clonedeep": "^4.5.0",
