+++ packages/stateful/proposal-module-adapter/adapters/CwdProposalSingle/common/components/NewProposal.tsx
@@ -1,34 +1,21 @@
-import { coins } from '@cosmjs/stargate'
-import { findAttribute } from '@cosmjs/stargate/build/logs'
 import { BookOutlined, FlagOutlined, Timelapse } from '@mui/icons-material'
 import { useWallet } from '@noahsaso/cosmodal'
-import { useCallback, useMemo, useState } from 'react'
-import { useFormContext } from 'react-hook-form'
+import { useMemo, useState } from 'react'
 import toast from 'react-hot-toast'
 import { useTranslation } from 'react-i18next'
-import {
-  constSelector,
-  useRecoilCallback,
-  useRecoilValue,
-  useSetRecoilState,
-} from 'recoil'
+import { useRecoilCallback, useRecoilValue } from 'recoil'
 
 import {
-  Cw20BaseSelectors,
   CwCoreV1Selectors,
   blockHeightSelector,
   blocksPerYearSelector,
   cosmWasmClientForChainSelector,
-  nativeDenomBalanceSelector,
-  refreshWalletBalancesIdAtom,
 } from '@dao-dao/state'
 import { useCachedLoadable, useDaoInfoContext } from '@dao-dao/stateless'
 import {
   Action,
   ActionKey,
   BaseNewProposalProps,
-  ContractVersion,
-  DepositInfoSelector,
   IProposalModuleAdapterCommonOptions,
   UseDefaults,
   UseTransformToCosmos,
@@ -36,23 +23,20 @@ import {
 import {
   convertExpirationToDate,
   dateToWdhms,
-  expirationExpired,
   processError,
 } from '@dao-dao/utils'
 
 import { useCoreActions } from '../../../../../actions'
-import {
-  Cw20BaseHooks,
-  useAwaitNextBlock,
-  useVotingModule,
-} from '../../../../../hooks'
+import { useVotingModule } from '../../../../../hooks'
 import { useVotingModuleAdapter } from '../../../../../voting-module-adapter'
-import { usePropose as useProposePrePropose } from '../../contracts/CwdPreProposeSingle.hooks'
 import { proposalSelector } from '../../contracts/CwdProposalSingle.common.recoil'
-import { usePropose as useProposeV2 } from '../../contracts/CwdProposalSingle.v2.hooks'
-import { usePropose as useProposeV1 } from '../../contracts/CwProposalSingle.v1.hooks'
 import { makeGetProposalInfo } from '../../functions'
-import { NewProposalData, NewProposalForm } from '../../types'
+import {
+  NewProposalData,
+  NewProposalForm,
+  PublishProposalOptions,
+  UsePublishProposal,
+} from '../../types'
 import {
   makeUseActions as makeUseProposalModuleActions,
   useProcessTQ,
@@ -61,20 +45,19 @@ import { NewProposal as StatelessNewProposal } from '../ui/NewProposal'
 
 export type NewProposalProps = BaseNewProposalProps<NewProposalForm> & {
   options: IProposalModuleAdapterCommonOptions
-  depositInfoSelector: DepositInfoSelector
+  usePublishProposal: UsePublishProposal
 }
 
 export const NewProposal = ({
   onCreateSuccess,
-  simulateMsgs,
   options,
-  depositInfoSelector,
+  usePublishProposal,
   ...props
 }: NewProposalProps) => {
   const { t } = useTranslation()
   const { imageUrl: daoImageUrl } = useDaoInfoContext()
   const { chainId, coreAddress, proposalModule } = options
-  const { connected, address: walletAddress } = useWallet()
+  const { connected } = useWallet()
   const { isMember = false } = useVotingModule(coreAddress, {
     fetchMembership: true,
   })
@@ -122,233 +105,38 @@ export const NewProposal = ({
     {}
   )
 
-  const formMethods = useFormContext<NewProposalForm>()
-
-  const depositInfo = useRecoilValue(depositInfoSelector)
-  const depositInfoCw20TokenAddress =
-    depositInfo?.denom && 'cw20' in depositInfo.denom
-      ? depositInfo.denom.cw20
-      : undefined
-  const depositInfoNativeTokenDenom =
-    depositInfo?.denom && 'native' in depositInfo.denom
-      ? depositInfo.denom.native
-      : undefined
-
   const blockHeightLoadable = useCachedLoadable(blockHeightSelector({}))
   const blockHeight =
     blockHeightLoadable.state === 'hasValue'
       ? blockHeightLoadable.contents
       : undefined
 
-  const requiredProposalDeposit = Number(depositInfo?.amount ?? '0')
-
-  // For checking allowance and increasing if necessary.
-  const cw20DepositTokenAllowanceResponseLoadable = useCachedLoadable(
-    depositInfoCw20TokenAddress && requiredProposalDeposit && walletAddress
-      ? Cw20BaseSelectors.allowanceSelector({
-          contractAddress: depositInfoCw20TokenAddress,
-          params: [
-            {
-              owner: walletAddress,
-              // If pre-propose address set, give that one deposit allowance
-              // instead of proposal module.
-              spender:
-                proposalModule.preProposeAddress || proposalModule.address,
-            },
-          ],
-        })
-      : constSelector(undefined)
-  )
-  const cw20DepositTokenAllowanceResponse =
-    cw20DepositTokenAllowanceResponseLoadable.state === 'hasValue'
-      ? cw20DepositTokenAllowanceResponseLoadable.contents
-      : undefined
-
-  const cw20DepositTokenBalanceLoadable = useCachedLoadable(
-    requiredProposalDeposit && walletAddress && depositInfoCw20TokenAddress
-      ? Cw20BaseSelectors.balanceSelector({
-          contractAddress: depositInfoCw20TokenAddress,
-          params: [{ address: walletAddress }],
-        })
-      : constSelector(undefined)
-  )
-  const cw20DepositTokenBalance =
-    cw20DepositTokenBalanceLoadable.state === 'hasValue'
-      ? cw20DepositTokenBalanceLoadable.contents
-      : undefined
-
-  const nativeDepositTokenBalanceLoadable = useCachedLoadable(
-    requiredProposalDeposit && walletAddress && depositInfoNativeTokenDenom
-      ? nativeDenomBalanceSelector({
-          walletAddress,
-          denom: depositInfoNativeTokenDenom,
-        })
-      : constSelector(undefined)
-  )
-  const nativeDepositTokenBalance =
-    nativeDepositTokenBalanceLoadable.state === 'hasValue'
-      ? nativeDepositTokenBalanceLoadable.contents
-      : undefined
-
-  // Whether or not deposit is can be paid or does not have to be.
-  const depositSatisfied =
-    requiredProposalDeposit === 0 ||
-    (cw20DepositTokenBalance &&
-      Number(cw20DepositTokenBalance.balance) >= requiredProposalDeposit) ||
-    (nativeDepositTokenBalance &&
-      Number(nativeDepositTokenBalance.amount) >= requiredProposalDeposit)
-
-  const setRefreshWalletBalancesId = useSetRecoilState(
-    refreshWalletBalancesIdAtom(walletAddress ?? '')
-  )
-  const refreshBalances = useCallback(
-    () => setRefreshWalletBalancesId((id) => id + 1),
-    [setRefreshWalletBalancesId]
-  )
-
   const processTQ = useProcessTQ()
-
-  const increaseCw20DepositAllowance = Cw20BaseHooks.useIncreaseAllowance({
-    contractAddress: depositInfoCw20TokenAddress ?? '',
-    sender: walletAddress ?? '',
-  })
-  const doProposeV1 = useProposeV1({
-    contractAddress: proposalModule.address,
-    sender: walletAddress ?? '',
-  })
-  const doProposeV2 = useProposeV2({
-    contractAddress: proposalModule.address,
-    sender: walletAddress ?? '',
-  })
-  const doProposePrePropose = useProposePrePropose({
-    contractAddress: proposalModule.preProposeAddress ?? '',
-    sender: walletAddress ?? '',
-  })
-
-  const awaitNextBlock = useAwaitNextBlock()
 
   const blocksPerYear = useRecoilValue(blocksPerYearSelector({}))
   const cosmWasmClient = useRecoilValue(cosmWasmClientForChainSelector(chainId))
+
+  const { publishProposal, depositUnsatisfied } = usePublishProposal()
+
   const createProposal = useRecoilCallback(
     ({ snapshot }) =>
-      async (newProposalData: NewProposalData) => {
-        if (!connected || blockHeight === undefined) {
+      async (
+        newProposalData: NewProposalData,
+        publishOptions?: PublishProposalOptions
+      ) => {
+        if (blockHeight === undefined) {
           toast.error(t('error.loadingData'))
           return
         }
 
         setLoading(true)
-
-        // Only simulate messages if any exist. Allow proposals without
-        // messages.
-        if (newProposalData.msgs.length > 0) {
-          try {
-            // Throws error if simulation fails, indicating invalid message.
-            await simulateMsgs(newProposalData.msgs)
-          } catch (err) {
-            console.error(err)
-            toast.error(
-              `${t('error.simulationFailedInvalidProposalActions')} ${
-                // Don't send to Sentry, but still format SDK errors nicely.
-                processError(err, { forceCapture: false })
-              }`
-            )
-            setLoading(false)
-            return
-          }
-        }
-
-        // Increase CW20 deposit token allowance if necessary.
-        if (requiredProposalDeposit && depositInfoCw20TokenAddress) {
-          // CW20 allowance response must be checked.
-          if (!cw20DepositTokenAllowanceResponse) {
-            toast.error(t('error.loadingData'))
-            return
-          }
-
-          const remainingAllowanceNeeded =
-            requiredProposalDeposit -
-            // If allowance expired, none.
-            (expirationExpired(
-              cw20DepositTokenAllowanceResponse.expires,
-              blockHeight
-            )
-              ? 0
-              : Number(cw20DepositTokenAllowanceResponse.allowance))
-
-          // Request to increase the contract's allowance for the proposal
-          // deposit if needed.
-          if (remainingAllowanceNeeded) {
-            try {
-              await increaseCw20DepositAllowance({
-                amount: remainingAllowanceNeeded.toString(),
-                spender:
-                  // If pre-propose address set, give that one deposit allowance
-                  // instead of proposal module.
-                  proposalModule.preProposeAddress || proposalModule.address,
-              })
-
-              // Allowances will not update until the next block has been added.
-              awaitNextBlock().then(refreshBalances)
-            } catch (err) {
-              console.error(err)
-              toast.error(
-                `Failed to increase allowance to pay proposal deposit: (${processError(
-                  err
-                )})`
-              )
-              setLoading(false)
-              return
-            }
-          }
-        }
-
-        // Native token deposit if exists.
-        const proposeFunds =
-          requiredProposalDeposit && depositInfoNativeTokenDenom
-            ? coins(requiredProposalDeposit, depositInfoNativeTokenDenom)
-            : undefined
-
         try {
-          let response
-          //! V1
-          if (proposalModule.version === ContractVersion.V1) {
-            response = await doProposeV1(
-              newProposalData,
-              'auto',
-              undefined,
-              proposeFunds
-            )
-            //! V2
-          } else {
-            response = proposalModule.preProposeAddress
-              ? await doProposePrePropose(
-                  {
-                    msg: {
-                      propose: newProposalData,
-                    },
-                  },
-                  'auto',
-                  undefined,
-                  proposeFunds
-                )
-              : await doProposeV2(
-                  newProposalData,
-                  'auto',
-                  undefined,
-                  proposeFunds
-                )
-          }
-
-          if (proposeFunds?.length) {
-            refreshBalances()
-          }
-
-          const proposalNumber = Number(
-            findAttribute(response.logs, 'wasm', 'proposal_id').value
+          const { proposalNumber, proposalId } = await publishProposal(
+            newProposalData,
+            publishOptions
           )
-          const proposalId = `${proposalModule.prefix}${proposalNumber}`
 
+          // Get proposal info to display card.
           const proposalInfo = await makeGetProposalInfo({
             ...options,
             proposalNumber,
@@ -381,8 +169,8 @@ export const NewProposal = ({
             proposalInfo
               ? {
                   id: proposalId,
-                  title: formMethods.getValues('title'),
-                  description: formMethods.getValues('description'),
+                  title: newProposalData.title,
+                  description: newProposalData.description,
                   info: [
                     {
                       Icon: BookOutlined,
@@ -412,8 +200,8 @@ export const NewProposal = ({
                 }
               : {
                   id: proposalId,
-                  title: formMethods.getValues('title'),
-                  description: formMethods.getValues('description'),
+                  title: newProposalData.title,
+                  description: newProposalData.description,
                   info: [],
                   dao: {
                     coreAddress,
@@ -429,29 +217,17 @@ export const NewProposal = ({
         }
       },
     [
-      connected,
       blockHeight,
-      requiredProposalDeposit,
-      depositInfoCw20TokenAddress,
-      depositInfoNativeTokenDenom,
-      t,
-      simulateMsgs,
-      cw20DepositTokenAllowanceResponse,
-      increaseCw20DepositAllowance,
-      proposalModule,
-      awaitNextBlock,
-      refreshBalances,
-      options,
-      cosmWasmClient,
       blocksPerYear,
-      processTQ,
-      onCreateSuccess,
-      formMethods,
       coreAddress,
+      cosmWasmClient,
       daoImageUrl,
-      doProposeV1,
-      doProposePrePropose,
-      doProposeV2,
+      onCreateSuccess,
+      options,
+      processTQ,
+      proposalModule.address,
+      publishProposal,
+      t,
     ]
   )
 
@@ -461,7 +237,7 @@ export const NewProposal = ({
       actionsWithData={actionsWithData}
       connected={connected}
       createProposal={createProposal}
-      depositUnsatisfied={!depositSatisfied}
+      depositUnsatisfied={depositUnsatisfied}
       isMember={isMember}
       isPaused={isPaused}
       loading={loading}
