+++ packages/stateful/proposal-module-adapter/adapters/CwdProposalSingle/common/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './makeUseActions'
 export * from './makeUseProfileNewProposalCardInfoLines'
+export * from './makeUsePublishProposal'
 export * from './useProcessTQ'
