+++ packages/stateful/proposal-module-adapter/adapters/CwdProposalSingle/index.tsx
@@ -10,6 +10,7 @@ import {
   makeDepositInfoSelector,
   makeUseActions,
   makeUseProfileNewProposalCardInfoLines,
+  makeUsePublishProposal,
   reverseProposalInfosSelector,
 } from './common'
 import {
@@ -60,6 +61,11 @@ export const CwdProposalSingleAdapter: ProposalModuleAdapter<
       preProposeAddress: options.proposalModule.preProposeAddress,
     })
 
+    const usePublishProposal = makeUsePublishProposal({
+      options,
+      depositInfoSelector,
+    })
+
     return {
       // Fields
       fields: {
@@ -97,8 +103,8 @@ export const CwdProposalSingleAdapter: ProposalModuleAdapter<
       components: {
         NewProposal: (props) => (
           <NewProposal
-            depositInfoSelector={depositInfoSelector}
             options={options}
+            usePublishProposal={usePublishProposal}
             {...props}
           />
         ),
